모두 Setting 네임스페이스를 사용하도록 수정
[/] using namespace Setting 모두 삭제 후 직접 네임스페이스를 통해 접근하도록 수정

+++ 소스/Client/texture.h
@@ -13,7 +13,7 @@ public:
 	void UpdateShaderVariable(const ComPtr<ID3D12GraphicsCommandList>& commandList, int index = 0);
 	void ReleaseUploadBuffer();
 
-	UINT GetTextureCount() const { return m_textures.size(); }
+	size_t GetTextureCount() const { return m_textures.size(); }
 
 protected:
 	ComPtr<ID3D12DescriptorHeap>				m_srvHeap;
